fix: Review comments: Remove DefaultParticipantIdentityResolver
